backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic>=2.0
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf
